[inductor] Refactor post_grad.py
[ghstack-poisoned]
